pyproject.toml
[tool.poetry]
name = "jtdub-general-ansible"
version = "0.1.0"
description = "Generic Ansible Galaxy Collections"
authors = ["James Williams <james.williams@packetgeek.net>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = "^3.8"
ansible-base = "^2.10.17"

[tool.poetry.dev-dependencies]
isort = "^5.12.0"
black = "^23.3.0"
pylint = "^2.17.3"
invoke = "^2.1.1"
bandit = "^1.7.5"
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
